Bunch of restructing and fixes.

+++ src/libs/disk.h
@@ -1,12 +1,18 @@
+//! \file disk.h
+//!
+//! Defines disk parameters.
+
 #ifndef __DISK__
 #define __DISK__
 
 #include "hi_types.h"
 
+
 struct SectorList
 {
-    unsigned int  sector;
+    BYTE  sector;
 };
+
 
 class Disk
 {
@@ -26,6 +32,7 @@ public:
     virtual BYTE maxSide(void) = 0;
     virtual BYTE numSides(void);
 
+    // sectors for a given track/side on heath's hard-sectored disks, it's always 10.
     virtual BYTE minSector(BYTE track,
                            BYTE side) = 0;
     virtual BYTE maxSector(BYTE track,
@@ -36,18 +43,20 @@ public:
     virtual BYTE tpi(void) = 0;
     virtual BYTE density(void) = 0;
 
-    virtual int  sectorBytes(BYTE side,
+    // number of bytes for the sector and the raw sector
+    virtual WORD sectorBytes(BYTE side,
                              BYTE track,
                              BYTE sector) = 0;
-    virtual int  sectorRawBytes(BYTE side,
+    virtual WORD sectorRawBytes(BYTE side,
                                 BYTE track,
                                 BYTE sector) = 0;
 
-    virtual int  trackBytes(BYTE side, 
+    virtual WORD trackBytes(BYTE side, 
                             BYTE track) = 0;
-    virtual int  trackRawBytes(BYTE side,
+    virtual WORD trackRawBytes(BYTE side,
                                BYTE track) = 0;
 
+    // for a given track number, the physical track on the 1.2M floppy
     virtual BYTE physicalTrack(BYTE track) = 0;
 
     virtual int readSector(BYTE *buffer,
@@ -57,14 +66,14 @@ public:
                            BYTE sector) = 0;
 
     virtual void genBestReadOrder(SectorList *out,
-                                  int         track,
-                                  int         side);
+                                  BYTE        track,
+                                  BYTE        side);
 
 private:
     virtual void halfTheSectors(SectorList **out,
-                                int          track,
-                                int          side,
-                                int          which_half); 
+                                BYTE         track,
+                                BYTE         side,
+                                BYTE         which_half); 
 
 };
